+++ tests/testthat/test-all-mailinglists.R
@@ -2,8 +2,6 @@ context("All mailing lists that the user has access to on Qualtrics")
 
 test_that("all_mailinglists returns a tbl_df with expected column names and types", {
 
-  qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
-
   vcr::use_cassette("all_mailinglists", {
     x <- all_mailinglists()
   })