+++ tests/testthat/test-survey-questions.R
@@ -1,9 +1,6 @@
-
 context("Get survey questions for a survey")
 
 test_that("survey_questions() makes a request with expected structure, and parses response", {
-
-  qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
 
   vcr::use_cassette("survey_questions", {
     x <- survey_questions("SV_3gbwq8aJgqPwQDP")
@@ -19,12 +16,17 @@ test_that("survey_questions() makes a request with expected structure, and parse
 
 })
 
-test_that("survey_questions() throws an error", {
+# Change credentials to
+qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
 
-  qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
+test_that("survey_questions() throws an error where URL & key are bad", {
 
   expect_error(
     qualtRics::survey_questions("1234"),
     "you may not have the\nrequired authorization"
   )
 })
+
+# Restore the credentials for other tests:
+qualtrics_api_credentials(api_key = holder_API, base_url = holder_URL)
+