+++ NEWS.md
@@ -1,8 +1,13 @@
 # qualtRics (development version)
 
+# qualtRics 3.1.4
+
 - Add `fetch_distributions()` to access distribution data for a specific survey thanks to @dsen6644 (#169)
 - Handle mailing list embedded data better thanks to @dsen6644 (#175)
 - Updated links to API documentation
+- Create unique column names for questions using `choiceId` thanks to @lyh970817 (#182)
+- Fix bug when `include_questions` only contains one QID thanks to @lyh970817 (#197)
+- Generate correct/updated column mapping for survey responses thanks to @jmobrien (#199). These column mappings are available as an attribute on survey results or via the new `extract_colmap()` function. 
 
 # qualtRics 3.1.3